[FE] FEAT: Mission Axios Daily, Custom Add

See merge request s07-webmobile2-sub2/S07P12B103!93

+++ frontend/eco2/src/components/dailyMission/missionList/dailyCustomMissionList.jsx
@@ -4,23 +4,41 @@ import { customMission } from "../../../store/mission/customMissionSlice";
 import { useNavigate } from "react-router-dom";
 import { useSelector, useDispatch } from "react-redux/es/exports";
 import { customDeleteMission } from "../../../store/mission/customMissionSlice";
+import { postMission } from "../../../store/mission/missionMainSlice";
 
 const DailyCustomMissionList = ({ id }) => {
   const [cos, setCos] = useState([]);
-  const customList = useSelector(state => state.custom);
+  const customList = useSelector((state) => state.custom);
   const dispatch = useDispatch();
   const naviGate = useNavigate();
+  const [list, setList] = useState([]);
 
   useEffect(() => {
-    dispatch(customMission({ id })).then(res => {
+    dispatch(customMission({ id })).then((res) => {
       setCos(res.payload.customMissionList);
     });
   }, []);
 
-  const onDelete = deleteId => {
-    dispatch(customDeleteMission({ id: deleteId }));
+  const onDelete = (deleteId) => {
+    if (window.confirm("커스텀 미션을 삭제하시겠습니까?")) {
+      dispatch(customDeleteMission({ id: deleteId }));
+      alert("삭제 완료!");
+      window.location.replace("/dailymissionDetail");
+    } else {
+      alert("더 좋은 미션들을 만들어 보아요 :)");
+    }
   };
 
+  const onCusMission = (postId) => {
+    dispatch(postMission({ id, dailyCustomMissionList: postId })).then((res) => {
+      if (res.payload.status === 200) {
+        console.log("postId ===> ", postId);
+        console.log("ID여기서 ===> ", id);
+      }
+    });
+  };
+
+  // console.log("CustomMissionList ID확인 용도 ===> ", cos);
   return (
     <div>
       {cos.length === 0 ? (
@@ -30,7 +48,7 @@ const DailyCustomMissionList = ({ id }) => {
         </div>
       ) : (
         <div>
-          {cos.map(it => (
+          {cos.map((it) => (
             <div key={it.id}>
               <span>{it.content}</span>
               <button
@@ -39,6 +57,13 @@ const DailyCustomMissionList = ({ id }) => {
                 }}
               >
                 삭제
+              </button>
+              <button
+                onClick={() => {
+                  onCusMission([it.id]);
+                }}
+              >
+                저장
               </button>
             </div>
           ))}
@@ -49,7 +74,11 @@ const DailyCustomMissionList = ({ id }) => {
         <i
           className={"fa-solid fa-circle-plus"}
           onClick={() => {
-            naviGate("/customPlus");
+            if (window.confirm("나만의 미션을 만들겠어요?")) {
+              naviGate("/customPlus");
+            } else {
+              alert("다음에는 미션을 만들어주세요!");
+            }
           }}
         ></i>
       </div>
